refactor: extract ResourceListItem component to separate file
- Move ResourceListItem component from ResourceListCourseContent.tsx to new ResourceListItem.tsx file
- Update import in ResourceDisplay.tsx to use new file location
- Remove circular import dependency and associated ESLint disable comment
- Clean up unused imports in ResourceListCourseContent.tsx
